feat: Implement functionality to end the game
The code changes in `BEndGameButton.java` implement the functionality to end the game when the button is clicked. This resolves the issue of the game not being able to be ended. The commit also includes the deletion of unused code and files related to the Binary Game.

Related user commits:
- Add new classes and files for Binary Game